backend/requirements.txt
Flask
flask-cors
textblob
mysql-connector-python
Werkzeug
